requirements.txt
# Core Django packages
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
django-filter==24.3
python-decouple==3.8
Pillow==10.4.0
requests==2.32.3

# Database support
mysqlclient==2.2.4  # MySQL support
psycopg2-binary==2.9.10  # PostgreSQL support
dj-database-url==2.1.0  # Database URL parsing

# Production server
gunicorn==21.2.0
whitenoise==6.6.0  # Static file serving

# Caching (Redis)
django-redis==5.4.0
redis==5.0.1

# Security and monitoring
django-ratelimit==4.1.0
sentry-sdk[django]==1.38.0  # Error tracking (optional)

# File handling and storage
boto3==1.34.0  # AWS S3 support (optional)
django-storages==1.14.2  # Cloud storage backends (optional)

# Development and testing (keep for migrations)
setuptools<81

# Health checks
django-health-check==3.17.0
